+++ R/create_test_assoc_qt_params.R
@@ -4,7 +4,7 @@
 #' to detect associations with quantitative traits
 #' (using \link{assoc_qt}) used for testing.
 #'
-#' @note This function is named after the \code{--assoc-qt} PLINK flag.
+#' @note This function is named after the \code{--assoc} PLINK flag.
 #' @inheritParams default_params_doc
 #' @author Richèl J.C. Bilderbeek
 #' @export