+++ tests/testthat/test-create_test_assoc_params.R
@@ -1,0 +1,3 @@
+test_that("use", {
+  expect_silent(create_test_assoc_params())
+})